requirements.txt
fastapi
uvicorn
bcrypt
pyjwt
pydantic[email]
sqlalchemy
psycopg2-binary
python-jose[cryptography]
slowapi
python-dotenv